# api-spec.yaml
...
openapi: "3.0.1"
info: 
  title: "Your API"
  version: "v1"
paths: 
  /api/Register: 
    post: 
      tags: 
      - "Auth"
      requestBody: 
        content: 
          application/json: 
            schema: 
              $ref: "#/components/schemas/RegiseterUserDto"
            example: 
              userName: "john.doe"
              email: "john.doe@example.com"
              password: "CapitalLetterNumberAndSymbol(JohnDoe4!)"
          text/json: 
            schema: 
              $ref: "#/components/schemas/RegiseterUserDto"
            example: 
              userName: "john.doe"
              email: "john.doe@example.com"
              password: "CapitalLetterNumberAndSymbol(JohnDoe4!)"
          application/*+json: 
            schema: 
              $ref: "#/components/schemas/RegiseterUserDto"
            example: 
              userName: "john.doe"
              email: "john.doe@example.com"
              password: "CapitalLetterNumberAndSymbol(JohnDoe4!)"
      responses: 
        201: 
          description: "Created"
        400: 
          description: "Bad Request"
        409: 
          description: "Conflict"
  /api/Login: 
    post: 
      tags: 
      - "Auth"
      requestBody: 
        content: 
          application/json: 
            schema: 
              $ref: "#/components/schemas/LoginUserDto"
            example: 
              userName: "john.doe"
              password: "CapitalLetterNumberAndSymbol(JohnDoe4!)"
          text/json: 
            schema: 
              $ref: "#/components/schemas/LoginUserDto"
            example: 
              userName: "john.doe"
              password: "CapitalLetterNumberAndSymbol(JohnDoe4!)"
          application/*+json: 
            schema: 
              $ref: "#/components/schemas/LoginUserDto"
            example: 
              userName: "john.doe"
              password: "CapitalLetterNumberAndSymbol(JohnDoe4!)"
      responses: 
        200: 
          description: "OK"
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/SuccessfulLoginUserDto"
              example: 
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy93cy8yMDA1LzA1L2lkZW50aXR5L2NsYWltcy9uYW1lIjoiYWRtaW4iLCJqdGkiOiJhYzY3MWI5NS00ODRhLTRmMDktOGNkZS04ZWI1N2RkZTlhMGUiLCJzdWIiOiI1Y2ZkMGZjZS0yNWQyLTRlZTQtYTk4ZC04ODdjNDI2OWVmMjYiLCJodHRwOi8vc2NoZW1hcy5taWNyb3NvZnQuY29tL3dzLzIwMDgvMDYvaWRlbnRpdHkvY2xhaW1zL3JvbGUiOlsiQWRtaW4iLCJVc2VyIl0sImV4cCI6MTczNDI3MzE0MSwiaXNzIjoiTGFpbWlzIiwiYXVkIjoiVHJ1c3RlZENsaWVudCJ9.Ev77_w9gQksd4SMAKmjHmvJrrrEP7JNjITdNK_Hs3T8"
        400: 
          description: "Bad Request"
        409: 
          description: "Conflict"
  /api/AccessToken: 
    post: 
      tags: 
      - "Auth"
      responses: 
        200: 
          description: "OK"
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/SuccessfulLoginUserDto"
              example: 
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy93cy8yMDA1LzA1L2lkZW50aXR5L2NsYWltcy9uYW1lIjoiYWRtaW4iLCJqdGkiOiJhYzY3MWI5NS00ODRhLTRmMDktOGNkZS04ZWI1N2RkZTlhMGUiLCJzdWIiOiI1Y2ZkMGZjZS0yNWQyLTRlZTQtYTk4ZC04ODdjNDI2OWVmMjYiLCJodHRwOi8vc2NoZW1hcy5taWNyb3NvZnQuY29tL3dzLzIwMDgvMDYvaWRlbnRpdHkvY2xhaW1zL3JvbGUiOlsiQWRtaW4iLCJVc2VyIl0sImV4cCI6MTczNDI3MzE0MSwiaXNzIjoiTGFpbWlzIiwiYXVkIjoiVHJ1c3RlZENsaWVudCJ9.Ev77_w9gQksd4SMAKmjHmvJrrrEP7JNjITdNK_Hs3T8"
        422: 
          description: "Unprocessable Content"
  /api/Logout: 
    post: 
      tags: 
      - "Auth"
      responses: 
        200: 
          description: "OK"
        422: 
          description: "Unprocessable Content"
...
components: 
  schemas: 
...
    LoginUserDto: 
      type: "object"
      properties: 
        userName: 
          type: "string"
          nullable: "true"
        password: 
          type: "string"
          nullable: "true"
      additionalProperties: "false"
...
    RegiseterUserDto: 
      type: "object"
      properties: 
        userName: 
          type: "string"
          nullable: "true"
        email: 
          type: "string"
          nullable: "true"
        password: 
          type: "string"
          nullable: "true"
      additionalProperties: "false"
...
    SuccessfulLoginUserDto: 
      type: "object"
      properties: 
        accessToken: 
          type: "string"
          nullable: "true"
      additionalProperties: "false"
...